requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
webdriver-manager>=4.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
pytz>=2024.1
openai>=1.0.0
gunicorn>=21.0.0
Werkzeug>=3.0.0
